+++ backend/src/main/java/com/onair/hearit/auth/application/OAuthService.java
@@ -1,5 +1,6 @@
 package com.onair.hearit.auth.application;
 
+import com.onair.hearit.auth.domain.OAuthProvider;
 import com.onair.hearit.auth.dto.request.OAuthUserInfo;
 
 public interface OAuthService {
