+++ backend/src/main/java/com/onair/hearit/auth/application/OAuthServiceRegistry.java
@@ -1,5 +1,6 @@
 package com.onair.hearit.auth.application;
 
+import com.onair.hearit.auth.domain.OAuthProvider;
 import com.onair.hearit.common.exception.custom.InvalidInputException;
 import java.util.List;
 import java.util.Map;
