+++ backend/src/main/java/com/onair/hearit/domain/Member.java
@@ -1,6 +1,6 @@
 package com.onair.hearit.domain;
 
-import com.onair.hearit.auth.application.OAuthProvider;
+import com.onair.hearit.auth.domain.OAuthProvider;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
