+++ backend/src/main/java/com/onair/hearit/infrastructure/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.onair.hearit.infrastructure;
 
-import com.onair.hearit.auth.application.OAuthProvider;
+import com.onair.hearit.auth.domain.OAuthProvider;
 import com.onair.hearit.domain.Member;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
